Generated by Spark: - Make the font for the ArticleContent a system default sans-serif font. - Move the table of contents to the left side of the page. - In the FFT preview component, please make the preview window persist when the cursor enters it while it is open. - The preview component needs to have the first sentence of the FFT article: "A fast Fourier transform (FFT) is an algorithm that computes the discrete Fourier transform (DFT) of a sequence, or its inverse (IDFT). A Fourier transform converts a signal from its original domain (often time or space) to a representation in the frequency domain and vice versa." - In line with that introduction, please add the hero image from the FFT article (attached). - below the introduction, include two tappable boxes, one labelled "Introduction" that includes another sentence from the FFT article: "Fast Fourier transforms are widely used for applications in engineering, music, science, and mathematics. The basic ideas were popularized in 1965, but some algorithms had been derived as early as 1805. In 1994, Gilbert Strang described the FFT as "the most important numerical algorithm of our lifetime", and it was included in Top 10 Algorithms of 20th Century by the IEEE magazine Computing in Science & Engineering.”
The second box will be labelled "Applications" and have a quote from the FFT article: “The FFT is used in digital recording, sampling, additive synthesis and pitch correction software.”

Clicking the introduction will bring the user to the FFT article in a different tab. Clicking the second box will open the FFT article in a new tab and scroll the user to the Introduction section. Clicking the third box will open the FFT article in a new tab and bring the user to the "Applications" section of the FFT article.